pyproject.toml
[tool.poetry]
name = "expense-app"
version = "0.1.0"
description = "Expense Application Backend"
authors = ["MAhmadAzhar <mahmadazhar@outlook.com>"]
readme = "README.md"
packages = [{include = "expense_app"}]

[tool.poetry.dependencies]
python = "3.10.9"
sqlmodel = "^0.0.8"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
